pyproject.toml
[project]
name = "car-rental-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "django>=5.2.6",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "dotenv>=0.9.9",
    "isort>=6.0.1",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.4.2",
    "pytest-django>=4.11.1",
    "python-dotenv>=1.1.1",
    "django-filter>=24.3",
    "django-storages[boto3]>=1.14.4",
    "Pillow>=10.4.0",
    "stripe>=12.5.1",
    "drf-spectacular>=0.28.0",
]
